move misplaced test
